Pass 'run_job_func' closure to _do_run_job()', rather than importing it.
Previously, "run_job" was imported from executors/base.py, and was run
by each exectuor when _do_submit_job() was called.

Now, in executors/base.py, 'run_job_func' is generated by calling
'generate_run_job_closure(job_submission_id)'. This function is
called in exectuors.py/base.py -> _do_prepare_submit_job(), which is
a wrapper around '_do_submit_job()'.

When 'generate_run_job_closure()' is called, it is passed a
"job_submission_id", which represents the job submission instance's
ID in the jobstore. This ID is bound to the inner-function "run_job()",
which creates a closure allowing "run_job()" access to the
job_submission's ID when it is later executed.

We eliminate the need to import "run_job" from base.py, and we pass this
closure (designated by 'run_job_func') to _do_submit_job(), where it is
able to be executed, allowing us to track the STATUS of the
job_submission (given that we now have local access to it's ID),
updating its status in the DB throughout its lifecylce.

Note that this is OS-agnostic - we don't want to get into the business
of tracking PIDs, etc... so we address this at the application-level
with some clever closure usage.

We also create the wrapper function "_do_prepare_submit_job()", which
adds the "job_submission" to the DB, tracks the job_submission's ID
in memory (self._instances), and then calls the old function
"_do_run_job()". More to come on the "job_submission" Table in
the next commit. (Keeping separate for organizational purposes).

*TODO: Still need to get clever with the Python3.5 coroutine
"run_job()".