Set running job_submissions to 'orphaned' on shutdown/start
Added update_job_submissions() Abstract Method to base.py, which
will be called to update job_submissions that satisfiy ``condition``,
and will set columns according to ``**kwargs``.

This function is called by jobstore.start and jobstore.shutdown to
set all job submissions that HAVE NOT COMPLETED to 'orphaned'. This
way we can track all jobs that never finished.

*Note depending on the type of scheduler, 'orphaned' status may not
properly describe job_submissions that are still 'running' when the
jobstore shutsdown. In most cases (that I can think of), these jobs
will stop when the scheduler stops (as they are processes, or threads
underneath the master APScheduler process). However if the jobstore
stops while the scheduler is still running, then as far as the jobstore
knows, these job submissions are 'orphaned'.