Syntax, variable names &run_job_func() controls job sub's state
1. Fixed lots of syntax errors

2. Updated code to meet PEP8 (majority, few outliers will address
before Pull Request)

3. Renamed prepare_submit_job() to prepare_job_submission()

4. Storing only 'job.func.__name__' in apscheduler_job_submissions store.

5. apscheduler_job_submission.status col name changed to 'state'

6. Create job submission table BEFORE updating job submissions
that are 'running' in the DB to 'orphaned' when we call jobstore.start()

7. run_job_func() now handles all updating of job_submission state
in the database, and in the jobstore._instances dictionary. This
was previously handled by exectuors.py -> _run_job_success() and
"_failure". Essentially, run_job_func() will set a job_submission
to "success" or "failure" in the database, and will also update
the jobstore._instances to remove the job from the list.**

8. Support coroutines (python3.5)