Replace run_job closure by passing job_submission_id as arg
The big change here is to remove the closure from exectuors/base.py.

This was initially meant to bind ``job_submission_id`` to the
``run_job()`` function, but we can, instead, pass this value
to ``_do_submit_job()``, which makes this available to the
callback function of the executor, which passes the argument
to the ``run_job()`` function. This function now is responsible
for UPDATE operations on the job_submissions collection in the
job store: tracking the job submission throughout its lifecycle.

Updated ``_do_submit_job()`` across all executors.

EVENT_JOB_EXECUTED & EVENT_JOB_ERROR now expose the job_submission_id
to the Public API.