Pass only values to ``run_job()``
Because ``run_job()`` is sometimes run as a separate process or thread,
we cannot pass it Scheduler objects, or Jobstore objects to update. We
can only send data in, and get events out.

This is fine - only we can't
distinguish between when a job is "submitted" to a Pool, and when it is
actually "executed"