Fixed the shutdown procedure of the Redis event broker
The client was not closed when the RedisEventBroker was initialized using RedisEventBroker.from_url().
Also, the pub-sub instance was not closed if the broker was shut down normally.
Additionally, when it was shut down due to an error, it used the deprecated .close() method to shut down the pub-sub instance.