Fix broken job lease extension
The `extend_job_leases` function was checking for `RunState.started` in a `while` loop,
however the coroutine is first executed by the (already running) job executors when the
scheduler is still `RunState.starting`. This caused the loop condition to immediately
evaluate to `False` and the function to return. This commit fixes that by checking for
both `RunState.starting` and `RunState.started` in `extend_job_leases`, similar to the
`cleanup` task.

With the `MemoryDataStore`, if a job ran longer than the lease *and* past the next
scheduled execution, then the exact same job would be executed again (along with a job
for the next schedule). The first execution cleans up when it finally finishes, but then
  the second execution fails in cleanup.

The issue primarily affects the `MemoryDataStore` because its `acquire_jobs` method
seems to acquire even expired jobs (which the scheduler re-executes) more aggressively
than the other data stores, which have different logic for discarding jobs / handling
task job slots.

This commit partially fixes #803, #834, #952, #967, and maybe others. I suspect there
are more bugs lurking that I haven't pinned down yet, likely related to the acquiring,
releasing, and perhaps cleanup logic.